partnerOrder jobs

+++ app/Models/PartnerOrder.php
@@ -252,11 +252,11 @@ class PartnerOrder extends Model
 
     public function getVersion()
     {
-        return $this->id > env('LAST_PARTNER_ORDER_ID_V1') ? 'v2' : 'v1';
+        return $this->id > (double)env('LAST_PARTNER_ORDER_ID_V1') ? 'v2' : 'v1';
     }
 
     public function getIsV2Attribute()
     {
-        return $this->id > env('LAST_PARTNER_ORDER_ID_V1');
+        return $this->id > (double)env('LAST_PARTNER_ORDER_ID_V1');
     }
 }
